requirements.txt
Flask
flask_sqlalchemy
flask_migrate
flask_login
flask_cors
jinja2
postgres
psycopg2-binary
python-dotenv
requests
requests_oauthlib
certifi
